refactor: organize files tree
