+++ app/src/main/java/com/example/tier_list_app/activities/home/HomeActivity.java
@@ -1,4 +1,4 @@
-package com.example.tier_list_app;
+package com.example.tier_list_app.activities.home;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -15,6 +15,9 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.tier_list_app.R;
+import com.example.tier_list_app.activities.tier_list.RegistryTierListActivity;
+import com.example.tier_list_app.activities.tier_list.TierListViewActivity;
 import com.example.tier_list_app.database.DBHelper;
 import com.example.tier_list_app.model.TierList;
 import com.example.tier_list_app.model.User;
@@ -99,7 +102,7 @@ public class HomeActivity extends AppCompatActivity {
                             @Override
                             public void onClick(View view) {
                                 TierList selectedTierList = arrayListTierList.get(position);
-                                Intent intent = new Intent(HomeActivity.this, TierListActivity.class);
+                                Intent intent = new Intent(HomeActivity.this, TierListViewActivity.class);
                                 intent.putExtra("chave_tier_list_id", selectedTierList.getId());
                                 intent.putExtra("tier_list_name", selectedTierList.getName());
                                 intent.putExtra("chave_usuario", selectedTierList.getUsername());
