+++ app/src/main/java/com/example/tier_list_app/activities/tier_list/TierListAdapter.java
@@ -1,4 +1,4 @@
-package com.example.tier_list_app;
+package com.example.tier_list_app.activities.tier_list;
 
 import android.view.LayoutInflater;
 import android.view.View;
@@ -9,6 +9,7 @@ import androidx.annotation.NonNull;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
+import com.example.tier_list_app.R;
 import com.example.tier_list_app.model.Tier;
 
 import java.util.List;
