+++ app/src/main/res/layout/activity_registy_tier.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context=".RegistryTierActivity">
+    tools:context=".activities.tier_list.RegistryTierActivity">
 
     <EditText
         android:id="@+id/edtName"
