+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="20dp"
-    tools:context=".SignUpUserActivity">
+    tools:context=".activities.auth.SignUpUserActivity">
     <EditText
         android:id="@+id/edtName"
         android:layout_width="match_parent"
