+++ app/src/main/res/layout/activity_tier_list.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".TierListActivity">
+    tools:context=".activities.tier_list.TierListViewActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/tierListRecyclerView"
